refactor: rename and replace about page components
