+++ src/components/AboutTabs/Description/Description.test.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable quotes*/
+
 import { render } from '@testing-library/react';
-import Description from './Description';
-import { studentAboutUsMock } from '../../../../__mocks__/AboutTabsMock';
+import Description from './Description.tsx';
+import { studentAboutUsMock } from '../../../__mocks__/AboutTabsMock.ts';
 
 describe('Test Description component', () => {
   test('Should render component with student information', () => {
@@ -8,8 +10,7 @@ describe('Test Description component', () => {
 
     expect(getByText('About Test')).toBeInTheDocument();
     expect(getByText('This is Test Name')).toBeInTheDocument();
-    // eslint-disable-next-line prettier/prettier
-    expect(getByText('Test\'s Contribution')).toBeInTheDocument();
+    expect(getByText("Test's Contribution")).toBeInTheDocument();
     expect(getByText('Contribution 1')).toBeInTheDocument();
     expect(getByText('Contribution 2')).toBeInTheDocument();
   });
@@ -19,8 +20,7 @@ describe('Test Description component', () => {
 
     expect(getByText('About Test')).toBeInTheDocument();
     expect(getByText('This is Test Name')).toBeInTheDocument();
-    // eslint-disable-next-line prettier/prettier
-    expect(getByText('Test\'s Contribution')).toBeInTheDocument();
+    expect(getByText("Test's Contribution")).toBeInTheDocument();
     expect(getByText('Contribution 1')).toBeInTheDocument();
     expect(getByText('Contribution 2')).toBeInTheDocument();
   });
