+++ src/components/AboutTabs/Description/description.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/common';
+@import '../../../styles/common/index';
 
 .root {
   max-width: 80%;
