+++ src/components/AboutTabs/Description/index.ts
@@ -1,0 +1,2 @@
+import Description from './Description.tsx';
+export { Description };
