+++ src/components/AboutTabs/Tab/index.ts
@@ -1,0 +1,2 @@
+import Tab from './Tab.tsx';
+export { Tab };
