+++ src/components/AboutTabs/Tab/tab.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/common';
+@import '../../../styles/common/index';
 
 .item {
   position: relative;
