+++ src/components/AboutTabs/aboutTabs.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import '../../styles/common/index';
 
 .root {
   margin-bottom: 2rem;
