+++ src/components/AboutTabs/index.ts
@@ -1,0 +1,2 @@
+import AboutTabs from './AboutTabs.tsx';
+export { AboutTabs };
