+++ src/components/AboutTabs/types/index.ts
