--- src/components/AboutUs/AboutTabs/Description/index.ts
@@ -1,2 +1,0 @@
-import Description from './Description';
-export { Description };
