--- src/components/AboutUs/AboutTabs/Tab/index.ts
@@ -1,2 +1,0 @@
-import Tab from './Tab';
-export { Tab };
