--- src/components/AboutUs/AboutTabs/index.ts
@@ -1,2 +1,0 @@
-import AboutTabs from './AboutTabs';
-export { AboutTabs };
