--- src/components/AboutUs/ContentBlock/index.ts
@@ -1,2 +1,0 @@
-import ContentBlock from './ContentBlock';
-export { ContentBlock };
