--- src/components/AboutUs/RSLogo/index.ts
@@ -1,2 +1,0 @@
-import RSLogo from './RSLogo';
-export { RSLogo };
