--- src/components/AboutUs/TeamProcess/index.ts
@@ -1,2 +1,0 @@
-import TeamProcess from './TeamProcess';
-export { TeamProcess };
