+++ src/components/TeamProcessList/ProcessItem/index.ts
@@ -1,0 +1,2 @@
+import ProcessItem from './ProcessItem.tsx';
+export { ProcessItem };
