+++ src/components/TeamProcessList/ProcessItem/processItem.module.scss
