+++ src/components/TeamProcessList/index.ts
@@ -1,0 +1,2 @@
+import TeamProcessList from './TeamProcessList.tsx';
+export { TeamProcessList };
