+++ src/components/TeamProcessList/teamProcessList.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import '../../styles/common/index';
 
 .root {
   padding: 5rem 5rem 7rem;
