+++ src/pages/AboutUs/RSLogo/index.ts
@@ -1,0 +1,2 @@
+import RSLogo from './RSLogo.tsx';
+export { RSLogo };
