+++ src/pages/AboutUs/RSLogo/rsLogo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/common';
+@import 'src/styles/common';
 
 .root {
   position: relative;
